requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
fastapi==0.115.14
uvicorn==0.35.0
httpx==0.28.1
Pillow==10.4.0 